Move AI locator classes into framework module, default to sonnet
