+++ framework/src/main/java/com/cucumberbddparallel/framework/ai/AiElementLocatorFactory.java
@@ -1,4 +1,4 @@
-package com.cucumberparallel.ai;
+package com.cucumberbddparallel.framework.ai;
 
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.SearchContext;
